Finished the draft for the article pages
Works fine exept there is no redirection from the confirmation of
creation to the page of the article.
We also need to add the links elements, the present date of creation and
decide wether we should add or not the option to let the user put a
banner for the article...
We will later have to add the saving without publishing option and for
that matter of fact, we will also have to make the cancel button work
(on this and on other creation page).
Last thing : the article page will have to be redone and so does the
creation page (for the latter, I will try to give it the same look as
the create event page but we will still have to decide on wether we
should keep certain options that are present on the page(view) but that
wern't included in the database )